FOODIES FAVE MANAGEMENT/GUIWithConsole/src/main/java/com/example/guiwithconsole/Customer.java
package com.example.guiwithconsole;

import javafx.beans.property.IntegerProperty;
import javafx.beans.property.SimpleIntegerProperty;
import javafx.beans.property.SimpleStringProperty;
import javafx.beans.property.StringProperty;

public class Customer {
    private String firstName;
    private String lastName;
    private int burgersRequired;

    public Customer(String firstName, String lastName, int burgersRequired) {
        this.firstName = firstName;
        this.lastName = lastName;
        this.burgersRequired = burgersRequired;
    }

    public String getFullName() {
        return firstName + " " + lastName;
    }

    public int getBurgersRequired() {
        return burgersRequired;
    }

    public StringProperty firstNameProperty() {
        return new SimpleStringProperty(firstName);
    }

    public StringProperty lastNameProperty() {
        return new SimpleStringProperty(lastName);
    }

    public IntegerProperty burgersRequiredProperty() {
        return new SimpleIntegerProperty(burgersRequired);
    }
}
